Fix compilation warnings on Win64
Fix compilation warnings on Win64, as shown below, by using proper long types
```
  gcompletion.c(47): warning C4028: formal parameter 3 different from declaration [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gconvert.c(105): warning C4133: 'function': incompatible types - from 'gsize *' to 'size_t *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gconvert.c(460): warning C4133: 'function': incompatible types - from 'size_t *' to 'gsize *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gdataset.c(552): warning C4311: 'type cast': pointer truncation from 'gpointer' to 'guint' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  giochannel.c(1238): warning C4133: 'function': incompatible types - from 'size_t *' to 'gsize *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gmem.c(662): warning C4311: 'type cast': pointer truncation from 'gpointer' to 'guint' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  ...
```

+++ graf2d/win32gdk/gdk/src/glib/gspawn-win32.c
@@ -188,7 +188,7 @@ read_data (GString     *str,
            GError     **error)
 {
   GIOError gioerror;
-  gint bytes;
+  gsize bytes;
   gchar buf[4096];
 
  again:
